File Backup and Version Update for functions.R and read_trees_data.R
Files were submitted for backup and version control.  I recently went and updated both the primary and source files to finalize the pruning and calculation of MS for partial trees.  There are issues with the current code pertaining to clades being reduced to 1 or 0 tips through the pruning process.  Specifically, the issue arrises when selecting for clades using the MRCA functions in the for loops used to fill the tree_list_half and tree_list_quarter objects.  I have attempted a work around which sets clades with 0 tips to have ages set to 0 whereas single tips are calculated normally.  Currently, the clades with 1 0r 0 tips are causing the MS functions to break during the bd.ms() function (line 49 of functions.R in the generation of the clade.crown.r object).